Game patch
Added controller for layer that allows mouse rotation to set the rotation of the player on Y axis.
Solved the lag problem at the start of game scene and tutorial scene. My async loading script was perfect, the problem was the huge amount of things done in Start() on functions. To solve this, I made a script, that runs in Edit Mode inside Unity, that is filling the requirments for other scripts, clearing the Start() when game plays (in build or in editor)